\name{pima_diabetes}
\alias{pima_diabetes}
\docType{data}
\title{Pima Indians diabetes data}
\description{
Binary diabetes endpoint with clinical predictors. Task type: classification.
}
\usage{
data(pima_diabetes)
}
\format{
A data frame with 532 rows and 9 columns.
}
\details{
Outcome column(s): \code{diabetes}.


}
\source{
\code{raw-selected-datasets/pima_diabetes/pima_diabetes.csv}. Smith et al. (1988), National Institute of Diabetes and Digestive and Kidney Diseases; UCI Machine Learning Repository.
}
\keyword{datasets}
